refactor: move schema related functions into a new file
