+++ core/tool_framework/registered_tool.py
@@ -2,12 +2,10 @@
 
 from __future__ import annotations
 
-import inspect
 from collections.abc import Callable, Iterable
 from copy import deepcopy
 from dataclasses import dataclass, field
-from types import NoneType
-from typing import Any, cast, get_args, get_origin, get_type_hints
+from typing import Any, cast, get_args
 
 from pydantic import BaseModel
 
@@ -15,6 +13,7 @@ from core.domain.types.evidence import EvidenceSource
 from core.domain.types.retrieval import RetrievalControls
 from core.domain.types.tools import ToolSurface
 from core.tool_framework.base import BaseTool, EvidenceType, SideEffectLevel, ToolMetadata
+from core.tool_framework.schema import _value_matches_schema, infer_input_schema, model_to_json_schema
 
 REGISTERED_TOOL_ATTR = "__opensre_registered_tool__"
 
@@ -46,135 +45,6 @@ def _normalize_surfaces(surfaces: Iterable[str] | None) -> tuple[ToolSurface, ..
 
     return tuple(normalized) or _DEFAULT_SURFACES
 
-
-def _strip_optional(annotation: Any) -> tuple[Any, bool]:
-    origin = get_origin(annotation)
-    if origin is None:
-        return annotation, False
-
-    args = tuple(arg for arg in get_args(annotation) if arg is not NoneType)
-    if len(args) != len(get_args(annotation)):
-        if len(args) == 1:
-            return args[0], True
-        return args, True
-
-    return annotation, False
-
-
-def _annotation_to_json_schema(annotation: Any) -> dict[str, Any]:
-    base_annotation, is_optional = _strip_optional(annotation)
-    origin = get_origin(base_annotation)
-
-    if base_annotation in (inspect.Signature.empty, Any):
-        schema: dict[str, Any] = {}
-    elif base_annotation is str:
-        schema = {"type": "string"}
-    elif base_annotation is int:
-        schema = {"type": "integer"}
-    elif base_annotation is float:
-        schema = {"type": "number"}
-    elif base_annotation is bool:
-        schema = {"type": "boolean"}
-    elif base_annotation is dict or origin is dict:
-        schema = {"type": "object"}
-    elif base_annotation is list or origin in (list, set, tuple):
-        schema = {"type": "array"}
-    else:
-        schema = {"type": "string"}
-
-    if is_optional:
-        schema["nullable"] = True
-    return schema
-
-
-def infer_input_schema(func: Callable[..., Any]) -> dict[str, Any]:
-    """Infer a minimal JSON schema from a function signature."""
-    properties: dict[str, Any] = {}
-    required: list[str] = []
-    type_hints = get_type_hints(func)
-
-    for param in inspect.signature(func).parameters.values():
-        if param.kind in (
-            inspect.Parameter.POSITIONAL_ONLY,
-            inspect.Parameter.VAR_POSITIONAL,
-            inspect.Parameter.VAR_KEYWORD,
-        ):
-            continue
-
-        if param.name.startswith("_"):
-            continue
-
-        resolved_annotation = type_hints.get(param.name, param.annotation)
-        schema = _annotation_to_json_schema(resolved_annotation)
-        properties[param.name] = schema
-
-        _, is_optional = _strip_optional(resolved_annotation)
-        if param.default is inspect.Signature.empty and not is_optional:
-            required.append(param.name)
-
-    return {
-        "type": "object",
-        "properties": properties,
-        "required": required,
-    }
-
-
-def model_to_json_schema(model: type[BaseModel]) -> dict[str, Any]:
-    """Convert a Pydantic model to a JSON object schema for tools."""
-    schema = model.model_json_schema()
-    if not isinstance(schema, dict):
-        return {"type": "object", "properties": {}, "required": [], "additionalProperties": False}
-    schema.setdefault("type", "object")
-    if schema.get("type") == "object":
-        schema.setdefault("properties", {})
-        schema.setdefault("required", [])
-        schema.setdefault("additionalProperties", False)
-    return schema
-
-
-def _json_type_matches(value: Any, schema_type: str) -> bool:
-    if schema_type == "string":
-        return isinstance(value, str)
-    if schema_type == "integer":
-        return isinstance(value, int) and not isinstance(value, bool)
-    if schema_type == "number":
-        return isinstance(value, (int, float)) and not isinstance(value, bool)
-    if schema_type == "boolean":
-        return isinstance(value, bool)
-    if schema_type == "array":
-        return isinstance(value, list)
-    if schema_type == "object":
-        return isinstance(value, dict)
-    return True
-
-
-def _value_matches_schema(value: Any, schema: dict[str, Any]) -> bool:
-    if value is None and bool(schema.get("nullable")):
-        return True
-
-    if "enum" in schema and value not in schema.get("enum", []):
-        return False
-
-    one_of = schema.get("oneOf")
-    if isinstance(one_of, list) and one_of:
-        return any(
-            isinstance(option, dict) and _value_matches_schema(value, option) for option in one_of
-        )
-
-    any_of = schema.get("anyOf")
-    if isinstance(any_of, list) and any_of:
-        return any(
-            isinstance(option, dict) and _value_matches_schema(value, option) for option in any_of
-        )
-
-    schema_type = schema.get("type")
-    if isinstance(schema_type, str):
-        return _json_type_matches(value, schema_type)
-    if isinstance(schema_type, list):
-        return any(
-            isinstance(item, str) and _json_type_matches(value, item) for item in schema_type
-        )
-    return True
 
 
 @dataclass
